docker: Notify docker not a requirement
Set a policy of not requiring docker as a development tool by giving
notice in the docker tool's README.

Docker's abstracts away a lot of effort when dealing with the
dependencies on a project.  The problem is when the project -relies- on
Docker, it adds an additional layer of abstraction and significant
resources to support it.